requirements.txt
appdirs==1.4.4
astroid==2.4.2
attrs==19.3.0
bleach==3.1.5
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
clipspy==0.3.3
colorama==0.4.3
docutils==0.16
Faker==4.1.2
fs==2.4.11
gevent==20.6.2
graph-theory==2020.5.6.39102
greenlet==0.4.16
h5py==2.10.0
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.0.1
isort==5.5.2
Keras==2.4.3
keyring==21.2.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.4.0
numpy==1.19.2
packaging==20.4
pandas==1.1.2
pkginfo==1.5.0.1
pluggy==0.13.1
py==1.9.0
pycparser==2.20
Pygments==2.6.1
pylint==2.6.0
pyNamespace==0.0.7
pyparsing==2.4.7
pytest==5.4.3
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.2
readme-renderer==26.0
requests==2.24.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
scipy==1.5.2
six==1.15.0
text-unidecode==1.3
toml==0.10.1
tqdm==4.47.0
twine==3.2.0
typed-ast==1.4.1
urllib3==1.25.9
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.12.1
zipp==3.1.0
zmq==0.0.0
zope.event==4.4
zope.interface==5.1.0
